measurement rest interface is added to handle measurement data
