+++ app/blueprints/rooms.py
@@ -1,9 +1,7 @@
 from flask import Blueprint, request
 from app.extensions import mongo
 from app.models.rooms import Room, RoomSchema
-from app.celery_tasks import long_task
 from app.models.device import Device, DeviceSchema
-from app.models.measurement import Measurement, MeasurementSchema
 from app.utils.responses import success_response, error_response
 from marshmallow import ValidationError 
 from app.utils.auth_utils import roles_required
@@ -13,7 +11,6 @@ room_schema = RoomSchema()
 rooms_schema = RoomSchema(many=True)
 device_schema = DeviceSchema()
 sensors_schema = DeviceSchema(many=True)
-measurement_schema = MeasurementSchema()
 
 
 @rooms_bp.route('/sensors', methods=['POST'])
@@ -106,31 +103,3 @@ def delete_room(room_id):
             return error_response("Failed to delete room", 500)
     except Exception as e:
         return error_response(str(e), 500)
-
-@rooms_bp.route('measurements/<string:device_id>', methods=['POST'])
-@roles_required('admin', 'user')
-def add_sensor_measurement(device_id):
-    """Adds a new measurement to a specific sensor in a room."""
-    try:
-        # 1. Retrieve the Sensor
-        sensor = Device.get_by_device_id(device_id)
-        print(sensor)
-        if not sensor:
-            return error_response("Sensor not found", 404)
-
-        # 2. Validate the new measurement data
-        data = request.get_json()
-        validated_data = measurement_schema.load(data)
-
-        # 3. Create a new Measurement object and associate it with the sensor
-        measurement = Measurement(**validated_data)
-        measurement.save()
-
-        # 4. Return success message
-        return success_response(message="Measurement added successfully",status_code=201)
-
-    except ValidationError as e:
-        return error_response(e.messages, 400)
-    except Exception as e:
-        return error_response(str(e), 500)
-
